move utils to same file
