--- packages/utils/src/iife.test.ts
@@ -1,11 +1,0 @@
-import { iife } from './iife';
-
-describe(iife.name, () => {
-  test('Should immediately invoke the given function and return its result.', () => {
-    const result = iife(() => {
-      return 1 + 1;
-    });
-
-    expect(result).toEqual(2);
-  });
-});
