--- packages/utils/src/to-escaped-string.test.ts
@@ -1,16 +1,0 @@
-import { toEscapedString } from './to-escaped-string';
-
-describe(toEscapedString.name, () => {
-  test('Should escape special characters.', () => {
-    expect(toEscapedString('foo.bar')).toBe('foo\\.bar');
-    expect(toEscapedString('foo[bar]')).toBe('foo\\[bar\\]');
-    expect(toEscapedString('foo(bar)')).toBe('foo\\(bar\\)');
-    expect(toEscapedString('foo{bar}')).toBe('foo\\{bar\\}');
-    expect(toEscapedString('foo?bar')).toBe('foo\\?bar');
-    expect(toEscapedString('foo+bar')).toBe('foo\\+bar');
-    expect(toEscapedString('foo*bar')).toBe('foo\\*bar');
-    expect(toEscapedString('foo^bar')).toBe('foo\\^bar');
-    expect(toEscapedString('foo$bar')).toBe('foo\\$bar');
-    expect(toEscapedString('foo|bar')).toBe('foo\\|bar');
-  });
-});
